requirements.txt
anyio==3.6.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
pydantic==1.10.2
PyJWT==2.5.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uuid==1.30
uvicorn==0.18.3
zipp==3.8.1
